medical-agent/pyproject.toml
[project]
name = "medical-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Guilherme Groff", email = "guilherme.pgb@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "debugpy>=1.8.17",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-mongodb>=0.7.1",
    "langchain-ollama>=1.0.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mysql-connector-python>=9.5.0",
    "pymysql>=1.1.2",
    "pypdf>=6.1.3",
]

[project.scripts]
medical-agent = "medical_agent:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
